\name{exampleChain1}
\alias{exampleChain1}
\docType{data}
\title{
Chain 1 from the OpenBUGS Pumps example
}
\description{
The coda object resulting from the Chain 1 using the model, data and inits
provided in the OpenBUGS Pumps example. 
}
\usage{data(exampleChain1)}
\source{
OpenBUGS Examples Vol 1
}
\references{
OpenBUGS Examples Vol 1
}
\keyword{datasets}
