\name{exampleChain2}
\alias{exampleChain2}
\docType{data}
\title{
Chain 2 from the OpenBUGS Pumps example
}
\description{
The coda object resulting from the Chain 2 using the model, data and inits
provided in the OpenBUGS Pumps example. 
}
\usage{data(exampleChain2)}
\source{
OpenBUGS Examples Vol 1
}
\references{
OpenBUGS Examples Vol 1
}
\keyword{datasets}
